Various changes to clean up makefiles a bit

+++ cpp/demo/Freeze/casino/Client.cpp
@@ -42,7 +42,7 @@ CasinoClient::run(int argc, char* argv[])
     //
     // Initialize pseudo-random number generator
     //
-    srand(IceUtil::Time::now().toMicroSeconds());
+    srand((unsigned int)IceUtil::Time::now().toMicroSeconds());
 
     cout << "Retrieve bank and players... " << flush;
